edge/requirements.txt
httpx>=0.27.0
supervision>=0.21.0
boxmot>=10.0.0
deepface>=0.0.89
insightface>=0.7.3
opencv-python-headless>=4.9.0
numpy>=1.26
ultralytics>=8.2.0
pyyaml>=6.0
